refactor(typescript): reorganize mechanisms/ to match Python directory layout
零行为改动，纯结构对齐 —— 把 TS mechanisms 从扁平布局重组成 Python 的
`mechanisms/{_base,_exact_base,_exact_permit_base,<chain>/<scheme>}` 结构：

旧文件 → 新位置：
  nativeExact.ts            → _exact_base/types.ts
  nativeExactEvm.ts         → evm/exact/client.ts
  nativeExactTron.ts        → tron/exact/client.ts
  exactEvm.ts               → evm/exact_permit/client.ts
  exact.ts                  → tron/exact_permit/client.ts
  exactGasfree.ts           → tron/exact_gasfree/client.ts
（同步移动相应的 *.test.ts 文件）

新增的目录索引文件（为后续 server/facilitator 端口铺路）：
  _base/index.ts               重新导出 3 个角色 ABC 接口（ClientMechanism、
                                ServerMechanism、FacilitatorMechanism）作为
                                mechanism-rooted 单一入口
  _exact_base/index.ts         exact scheme 共享 types
  _exact_permit_base/index.ts  exact_permit scheme 共享底座（待填充）
  evm/exact/index.ts           barrel
  evm/exact_permit/index.ts    barrel
  tron/exact/index.ts          barrel
  tron/exact_permit/index.ts   barrel
  tron/exact_gasfree/index.ts  barrel

主 mechanisms/index.ts 维持原有公共导出表（back-compat），客户代码无需改动。

修复所有 vi.mock 和 import 相对路径（从 1 层深的 'mechanisms/' 调整到
2 层深的 'mechanisms/<chain>/<scheme>/'）。

测试 144/144 全过，零回归。
