+++ typescript/packages/x402/src/mechanisms/_exact_base/types.ts
@@ -5,10 +5,10 @@
  * without going through a PaymentPermit contract.
  */
 
-import type { Hex } from '../address.js';
-import type { TransferAuthorization } from '../types/payment.js';
+import type { Hex } from '../../address.js';
+import type { TransferAuthorization } from '../../types/payment.js';
 
-export type { TransferAuthorization } from '../types/payment.js';
+export type { TransferAuthorization } from '../../types/payment.js';
 
 /** Scheme name */
 export const SCHEME_EXACT = 'exact';
