+++ typescript/packages/x402/src/mechanisms/evm/exact/index.ts
@@ -1,0 +1,1 @@
+export { ExactEvmClientMechanism } from './client.js';
