+++ typescript/packages/x402/src/mechanisms/evm/exact_permit/index.ts
@@ -1,0 +1,1 @@
+export { ExactPermitEvmClientMechanism } from './client.js';
