+++ typescript/packages/x402/src/mechanisms/tron/exact/index.ts
@@ -1,0 +1,1 @@
+export { ExactTronClientMechanism } from './client.js';
