+++ typescript/packages/x402/src/mechanisms/tron/exact_gasfree/index.ts
@@ -1,0 +1,1 @@
+export { ExactGasFreeClientMechanism } from './client.js';
