+++ typescript/packages/x402/src/mechanisms/tron/exact_permit/client.ts
@@ -1,5 +1,7 @@
 /**
- * ExactPermitEvmClientMechanism - EVM client mechanism for "exact_permit" payment scheme
+ * ExactPermitTronClientMechanism - TRON client mechanism for "exact_permit" payment scheme
+ *
+ * Uses TIP-712 (TRON's EIP-712 implementation) for signing PaymentPermit.
  */
 
 import type {
@@ -9,25 +11,25 @@ import type {
   PaymentPayload,
   PaymentPermit,
   PaymentPermitContext,
-} from '../index.js';
+} from '../../../index.js';
 import {
   KIND_MAP,
   PAYMENT_PERMIT_TYPES,
   PAYMENT_PERMIT_PRIMARY_TYPE,
   getChainId,
   getPaymentPermitAddress,
-  EVM_ZERO_ADDRESS,
-  EvmAddressConverter,
-  ZERO_ADDRESS_HEX,
+  TronAddressConverter,
+  TRON_ZERO_ADDRESS,
+  paymentIdToBytes,
   PermitValidationError,
-} from '../index.js';
+} from '../../../index.js';
 
 /**
- * EVM client mechanism for "exact_permit" payment scheme
+ * TRON client mechanism for "exact_permit" payment scheme
  */
-export class ExactPermitEvmClientMechanism implements ClientMechanism {
+export class ExactPermitTronClientMechanism implements ClientMechanism {
   private signer: ClientSigner;
-  private addressConverter = new EvmAddressConverter();
+  private addressConverter = new TronAddressConverter();
 
   constructor(signer: ClientSigner) {
     this.signer = signer;
@@ -52,7 +54,7 @@ export class ExactPermitEvmClientMechanism implements ClientMechanism {
     }
 
     const buyerAddress = this.signer.getAddress();
-    const zeroAddress = EVM_ZERO_ADDRESS;
+    const zeroAddress = TRON_ZERO_ADDRESS; // Use TRON zero address for consistency with Python
 
     const permit: PaymentPermit = {
       meta: {
@@ -83,7 +85,8 @@ export class ExactPermitEvmClientMechanism implements ClientMechanism {
       requirements.network
     );
 
-    // Build EIP-712 domain
+    // Build EIP-712 domain (no version field per contract spec)
+    // Note: domain name is "PaymentPermit", not "PaymentPermitDetails"
     const permitAddress = getPaymentPermitAddress(requirements.network);
     const domain = {
       name: 'PaymentPermit',
@@ -91,11 +94,15 @@ export class ExactPermitEvmClientMechanism implements ClientMechanism {
       verifyingContract: this.addressConverter.toEvmFormat(permitAddress),
     };
 
-    // Convert permit to EIP-712 compatible format
+    // Convert permit to EIP-712 compatible format:
+    // 1. kind: string -> uint8 (number)
+    // 2. paymentId: keep as hex string (TronWeb expects hex for bytes16)
+    // 3. uint256 values: string -> BigInt (for proper EIP-712 encoding)
+    // 4. All addresses: TRON Base58 -> EVM hex
     const permitForSigning = {
       meta: {
         kind: KIND_MAP[permit.meta.kind],
-        paymentId: permit.meta.paymentId,
+        paymentId: permit.meta.paymentId,  // Keep as hex string
         nonce: BigInt(permit.meta.nonce),
         validAfter: permit.meta.validAfter,
         validBefore: permit.meta.validBefore,
@@ -113,12 +120,26 @@ export class ExactPermitEvmClientMechanism implements ClientMechanism {
       },
     };
 
+    // Debug: log exact message being signed
+    console.log('[SIGN] Domain:', JSON.stringify(domain));
+    console.log('[SIGN] Message:', JSON.stringify(permitForSigning, (key, value) => {
+      if (value instanceof Uint8Array) {
+        return '0x' + Array.from(value).map(b => b.toString(16).padStart(2, '0')).join('');
+      }
+      if (typeof value === 'bigint') {
+        return value.toString();
+      }
+      return value;
+    }));
+
     const signature = await this.signer.signTypedData(
       domain,
       PAYMENT_PERMIT_TYPES,
       permitForSigning as unknown as Record<string, unknown>,
       PAYMENT_PERMIT_PRIMARY_TYPE
     );
+
+    console.log('[SIGN] Signature:', signature);
 
     return {
       x402Version: 2,
