+++ typescript/packages/x402/src/mechanisms/tron/exact_permit/index.ts
@@ -1,0 +1,1 @@
+export { ExactPermitTronClientMechanism } from './client.js';
